refactor navComp
